% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bduk_bq.R
\name{bduk_bq}
\alias{bduk_bq}
\title{Runs a query in BigQuery and loads the data into R.}
\usage{
bduk_bq(sql, project, key = NULL, keypath = NULL)
}
\arguments{
\item{sql}{SQL code to be queried}

\item{project}{The name of the bigquery project that the data is stored within. Unless otherwise specified, the key will be the name of this concatenated with "_bigquery.json".}

\item{key}{The name of the BigQuery key used for billing}

\item{keypath}{The path to the folder that contains the key. By default, this will be the current directory, and it is recommended the key is stored here.}
}
\value{
A data table containing the query results.
}
\description{
This function loads a BigQuery into a data table.
This requires both a billing string and a json key to be installed.
The json key must be in the same directory/project as you are working within.
}
\examples{
bduk_bq(
 sql="SELECT pcds,Rurality FROM `dcms-datalake-staging.GEO_ONS.ONS_RURALITY` LIMIT 1 ",
 project="dcms-datalake-staging"
)
bduk_bq(
 sql="SELECT pcds,Rurality FROM `dcms-datalake-staging.GEO_ONS.ONS_RURALITY` LIMIT 1 ",
 project="dcms-datalake-staging",
 key="dcms-datalake-staging_bigquery.json",
 keypath="/home/dcms/keys"
)
}
