% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearGroups.R
\name{clearGroups}
\alias{clearGroups}
\title{Function to clear all groups from a leaflet object from within a shiny application.}
\usage{
clearGroups(map, ...)
}
\arguments{
\item{map}{a string containing the name of the map}

\item{...}{a comma seperated set of strings that contain the group names to clear
See README for example useage.}
}
\description{
Function to clear all groups from a leaflet object from within a shiny application.
}
