% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geojson_to_sf.R
\name{geojson_to_sf}
\alias{geojson_to_sf}
\title{Takes a dataframe or tibble that contains a geojson column and outputs this as a dataframe or tibble with simple features geometry.}
\usage{
geojson_to_sf(data, geojson_colname)
}
\arguments{
\item{data}{a dataframe or tibble containing a geojson column}

\item{geojson_colname}{a string representing the geojson column name}
}
\value{
A simple feature data frame or tibble.
}
\description{
Takes a dataframe or tibble that contains a geojson column and outputs this as a dataframe or tibble with simple features geometry.
}
\examples{
geojson_to_sf(
data=bduk_bq(
sql="SELECT * FROM dcms-datalake-staging.GEO_ONS.shp_LA LIMIT 1",
project="dcms-datalake-staging",
keypath="/home/dcms/keys"),
geojson_colname = "geom"
)
}
