% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Shorthand for the opposite of in written in a more intuitive format. Returns true if the item is not in what it is being compared to, and false if it is in it.}
\usage{
... \%notin\% NA
}
\arguments{
\item{data}{any data type}
}
\value{
logical
}
\description{
Shorthand for the opposite of in written in a more intuitive format. Returns true if the item is not in what it is being compared to, and false if it is in it.
}
\examples{
2 \%notin\% c(1,3)
"b" \%notin\% c("a","c")
}
