% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_integer64.R
\name{is_integer64}
\alias{is_integer64}
\title{Return true or false if the object is integer64 type}
\usage{
is_integer64(x)
}
\arguments{
\item{x}{is an object}
}
\value{
boolean
}
\description{
Return true or false if the object is integer64 type
}
\examples{
is_integer64(100)
}
