% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_postcode.R
\name{is_postcode}
\alias{is_postcode}
\title{Function to check if a string is a valid postcode regardless of it's formatting. With or without a space is permitted.}
\usage{
is_postcode(Postcode)
}
\arguments{
\item{Postcode}{string}
}
\value{
Logical
}
\description{
Function to check if a string is a valid postcode regardless of it's formatting. With or without a space is permitted.
}
\examples{
is_postcode("SW1a2nP")
is_postcode("SW1a 2nP")
}
