% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_uprn.R
\name{is_uprn}
\alias{is_uprn}
\title{Function to check is either a numeric or character variable is in the format of a valid UPRN, and not not contain any letters or punctuation. This function will faill if a numeric with a decimal followed by solely zeros is entered.}
\usage{
is_uprn(uprns)
}
\arguments{
\item{string}{to be checked.}
}
\value{
TRUE or FALSE.
}
\description{
Function to check is either a numeric or character variable is in the format of a valid UPRN, and not not contain any letters or punctuation. This function will faill if a numeric with a decimal followed by solely zeros is entered.
}
\examples{
is_uprn(1)
is_uprn(999999999999)
is_uprn("1")
is_uprn("999999999999")
}
