% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_postcode.R
\name{pretty_postcode}
\alias{pretty_postcode}
\title{Turn a string into the 'prettty' format of postcode 'XX(X)(X)YXXX' where Y is a specified seperator with a spacer as default in either upper or lower case}
\usage{
pretty_postcode(Postcode, sep = " ", uppercase = TRUE)
}
\arguments{
\item{Postcode}{string}

\item{sep}{string}

\item{uppercase}{logical}
}
\value{
A string
}
\description{
Turn a string into the 'prettty' format of postcode 'XX(X)(X)YXXX' where Y is a specified seperator with a spacer as default in either upper or lower case
}
\examples{
pretty_postcode("SW1a2nP")
pretty_postcode("SW1a2nP", sep="")
pretty_postcode("SW1a2nP", sep=".")
pretty_postcode("SW1a2nP", sep=".", uppercase=FALSE)
}
