% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radioTooltip.R
\name{radioTooltip}
\alias{radioTooltip}
\title{Tooltip that only applies a pop up tooltip to specific selections of a shiny radio button.}
\usage{
radioTooltip(
  id,
  choice,
  title,
  placement = "bottom",
  trigger = "hover",
  options = NULL
)
}
\arguments{
\item{id}{The ID of the radio button the tooltip should appear on.}

\item{choice}{The choice within the radio buttons that the tooltip should appear on.}

\item{title}{The tooltip message as a string.}

\item{placement}{Where the popover should appear relative to its target
(\code{top}, \code{bottom}, \code{left}, or \code{right}). Defaults to \code{"bottom"}.}

\item{trigger}{What action should cause the popover to appear? (\code{hover},
\code{focus}, \code{click}, or \code{manual}). Defaults to \code{"hover"}.}

\item{options}{A named list of additional options to be set on the popover.

See README for example usage.}
}
\description{
Tooltip that only applies a pop up tooltip to specific selections of a shiny radio button.
}
