% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme.R
\name{readme}
\alias{readme}
\title{Shorthand to create a README file using the template stored within this package.}
\usage{
readme()
}
\value{
README.Rmd document.
}
\description{
Shorthand to create a README file using the template stored within this package.
}
\examples{
readme()

}
